Refactor: move core::forex.rs into its own directory module and separate things inside into their own files module
