+++ pfm-core/src/forex/mod.rs
@@ -1,0 +1,12 @@
+pub mod currency;
+pub use currency::Currency;
+
+pub mod entity;
+
+pub mod interface;
+pub use interface::{ForexError, ForexResult};
+
+pub mod money;
+pub use money::Money;
+
+pub mod service;
