+++ pfm-core/src/forex_impl/open_exchange_api.rs
@@ -7,9 +7,11 @@
 
 use std::str::FromStr;
 
-use crate::forex::ForexError::{self, OpenExchangeAPIError};
 use crate::forex::{
-    Currency, ForexHistoricalRates, ForexRates, HistoricalRates, RatesData, RatesResponse,
+    entity::{HistoricalRates, RatesData, RatesResponse},
+    interface::{ForexHistoricalRates, ForexRates},
+    Currency,
+    ForexError::{self, OpenExchangeAPIError},
 };
 use anyhow::anyhow;
 use async_trait::async_trait;
@@ -83,7 +85,7 @@ pub struct Rates {
     pub xpt: Decimal,
 }
 
-impl TryFrom<Response> for RatesResponse<crate::forex::Rates> {
+impl TryFrom<Response> for RatesResponse<crate::forex::entity::Rates> {
     type Error = ForexError;
 
     fn try_from(value: Response) -> Result<Self, Self::Error> {
@@ -119,7 +121,7 @@ impl TryFrom<Response> for RatesResponse<crate::forex::Rates> {
             ))
         })?;
 
-        let ret = crate::forex::Rates {
+        let ret = crate::forex::entity::Rates {
             latest_update: date,
             base,
             rates,
@@ -165,7 +167,7 @@ impl TryFrom<Response> for RatesResponse<HistoricalRates> {
             ))
         })?;
 
-        let ret = crate::forex::HistoricalRates { base, date, rates };
+        let ret = crate::forex::entity::HistoricalRates { base, date, rates };
 
         Ok(RatesResponse::new(SOURCE.into(), ret))
     }
@@ -191,7 +193,7 @@ impl ForexRates for Api {
     async fn rates(
         &self,
         base: Currency,
-    ) -> crate::forex::ForexResult<RatesResponse<crate::forex::Rates>> {
+    ) -> crate::forex::ForexResult<RatesResponse<crate::forex::entity::Rates>> {
         let params = [
             ("app_id", self.key),
             ("base", base.code()),
@@ -240,7 +242,7 @@ impl ForexHistoricalRates for Api {
         &self,
         date: chrono::DateTime<chrono::Utc>,
         base: Currency,
-    ) -> crate::forex::ForexResult<RatesResponse<crate::forex::HistoricalRates>> {
+    ) -> crate::forex::ForexResult<RatesResponse<crate::forex::entity::HistoricalRates>> {
         let yyyymmdd = date.format("%Y-%m-%d").to_string();
         let endpoint = HISTORICAL_ENDPOINT.replace(":date", yyyymmdd.as_str());
         let params = [
